integrate init.rs into session.rs
